MARS-1056 Remove create User workflow
